requirements.txt
aioredis==1.3.1
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.5
certifi==2021.10.8
cffi==1.15.0
channels==3.0.3
channels-redis==3.2.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Django==3.2
django-celery-beat==2.2.0
django-ckeditor==6.2.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-globals==0.3.2
django-js-asset==1.2.2
django-modeladmin-reorder==0.3.1
django-timezone-field==4.2.1
django-tinymce==3.2.0
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.1.0
fcm-django==0.3.7
ffmpy==0.3.0
filelock==3.4.0
flake8==4.0.1
gevent==21.8.0
greenlet==1.1.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.2.0
incremental==21.3.0
infi.clickhouse-orm==2.1.1
inflection==0.5.1
iso8601==1.0.2
kombu==5.2.2
lml==0.1.0
loguru==0.5.3
mccabe==0.6.1
msgpack==1.0.2
openpyxl==3.0.9
phonenumbers==8.12.37
Pillow==8.1.2
prompt-toolkit==3.0.22
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyexcel-io==0.6.5
pyexcel-xlsx==0.6.0
PyExifTool==0.1.1
pyfcm==1.4.9
pyflakes==2.4.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.15.0
pytils==0.3
pytz==2021.1
redis==3.5.3
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
transliterate==1.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==4.0.0
uritools==3.0.2
urlextract==1.4.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
win32-setctime==1.0.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
